Always assume HTTP path is from the root
This doesn't match the RFCs, but if we don't assume, it makes it a bit more annoying every time this library is used. I'm willing to sacrifice a tiny bit of functionality that may be used 0.000000001% of the time to make it better and easier to use the rest of the time.